Move tests to deprecated methods to separate class
